convert value to unsigned long to prevent overflow

+++ WaterManager.cpp
@@ -12,12 +12,12 @@ WaterManager::WaterManager() {
   superStateMainOn = new ValveState(valveMain, 0, "mainOn");
 
   stateIdle = new DurationState(0, "areasIdle", superStateMainIdle);
-  stateWarn = new ValveState(valveArea1, DURATION_WARN_SEC * 1000, "warn", superStateMainOn);
-  stateWaitBefore = new DurationState(DURATION_WAIT_BEFORE_SEC * 1000, "areasIdle", superStateMainIdle);
-  stateAutomatic1 = new ValveState(valveArea1, DURATION_AUTOMATIC1_SEC * 1000, "area1", superStateMainOn);
-  stateAutomatic2 = new ValveState(valveArea2, DURATION_AUTOMATIC2_SEC * 1000, "area2", superStateMainOn);
-//  stateAutomatic3 = new ValveState(valveArea3, DURATION_AUTOMATIC3_SEC * 1000, "area3", superStateMainOn);
-  stateManual = new ValveState(valveArea1, DURATION_MANUAL_SEC * 1000, "manual", superStateMainOn);
+  stateWarn = new ValveState(valveArea1, (unsigned long) DURATION_WARN_SEC * 1000, "warn", superStateMainOn);
+  stateWaitBefore = new DurationState((unsigned long) DURATION_WAIT_BEFORE_SEC * 1000, "areasIdle", superStateMainIdle);
+  stateAutomatic1 = new ValveState(valveArea1, (unsigned long) DURATION_AUTOMATIC1_SEC * 1000, "area1", superStateMainOn);
+  stateAutomatic2 = new ValveState(valveArea2, (unsigned long) DURATION_AUTOMATIC2_SEC * 1000, "area2", superStateMainOn);
+//  stateAutomatic3 = new ValveState(valveArea3, (unsigned long) DURATION_AUTOMATIC3_SEC * 1000, "area3", superStateMainOn);
+  stateManual = new ValveState(valveArea1, (unsigned long) DURATION_MANUAL_SEC * 1000, "manual", superStateMainOn);
 
   stateWarn->nextState = stateWaitBefore;
   stateWaitBefore->nextState = stateAutomatic1;
